rename: model 패키지 세분화하여 분리
