+++ src/main/java/lotto/dto/LottoTotalReturnDTO.java
@@ -2,7 +2,7 @@ package lotto.dto;
 
 import java.math.BigDecimal;
 import java.text.DecimalFormat;
-import lotto.model.LottoRankResult;
+import lotto.model.rank.LottoRankResult;
 
 public record LottoTotalReturnDTO(LottoRankResult result, int countOfLotto) {
     public String getTotalReturn() {
