--- src/main/java/lotto/model/LottoGenerator.java
@@ -1,5 +1,0 @@
-package lotto.model;
-
-public interface LottoGenerator {
-    Lotto generate();
-}
