+++ src/main/java/lotto/model/trade/LottoGenerator.java
@@ -1,0 +1,7 @@
+package lotto.model.trade;
+
+import lotto.model.lotto.Lotto;
+
+public interface LottoGenerator {
+    Lotto generate();
+}
