+++ src/main/java/lotto/model/trade/LottoSeller.java
@@ -1,8 +1,10 @@
-package lotto.model;
+package lotto.model.trade;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.model.lotto.Lotto;
+import lotto.model.lotto.Lottos;
 
 public class LottoSeller {
     private final LottoGenerator lottoGenerator;
