+++ src/main/java/lotto/service/LottoGameService.java
@@ -6,13 +6,13 @@ import java.util.stream.IntStream;
 import lotto.dto.LottoNumbersDTO;
 import lotto.dto.LottoRankResultDTO;
 import lotto.dto.LottoTotalReturnDTO;
-import lotto.model.LottoRank;
-import lotto.model.LottoRankResult;
-import lotto.model.Lotto;
-import lotto.model.LottoBuyer;
-import lotto.model.LottoSeller;
-import lotto.model.Lottos;
-import lotto.model.WinningLotto;
+import lotto.model.rank.LottoRank;
+import lotto.model.rank.LottoRankResult;
+import lotto.model.lotto.Lotto;
+import lotto.model.trade.LottoBuyer;
+import lotto.model.trade.LottoSeller;
+import lotto.model.lotto.Lottos;
+import lotto.model.rank.WinningLotto;
 import lotto.view.InputView;
 import lotto.view.OutputView;
