+++ src/test/java/lotto/model/AutoLottoGeneratorTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import lotto.ApplicationConfig;
+import lotto.model.lotto.Lotto;
+import lotto.model.trade.LottoGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
