+++ src/test/java/lotto/model/LottoBuyerTest.java
@@ -7,6 +7,7 @@ import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import lotto.ApplicationConfig;
+import lotto.model.trade.LottoBuyer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
