+++ src/test/java/lotto/model/LottoRankResultTest.java
@@ -7,6 +7,11 @@ import java.text.DecimalFormat;
 import java.util.List;
 import java.util.stream.IntStream;
 import lotto.ApplicationConfig;
+import lotto.model.lotto.Lotto;
+import lotto.model.lotto.Lottos;
+import lotto.model.rank.LottoRank;
+import lotto.model.rank.LottoRankResult;
+import lotto.model.rank.WinningLotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
