+++ src/test/java/lotto/model/LottoRankTest.java
@@ -3,6 +3,7 @@ package lotto.model;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.math.BigInteger;
+import lotto.model.rank.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
