+++ src/test/java/lotto/model/LottoSellerTest.java
@@ -3,6 +3,8 @@ package lotto.model;
 import static org.assertj.core.api.Assertions.*;
 
 import lotto.ApplicationConfig;
+import lotto.model.lotto.Lottos;
+import lotto.model.trade.LottoSeller;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
